feat(Calendario)  - calendario implementado com odas funcionalidades.

+++ Selecao/Selecao-Service/src/main/java/br/com/crescer/selecao/service/repository/CandidatoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 /**
  * @author Murillo
  */
-public interface CandidatoRepository extends PagingAndSortingRepository<Candidato, Long> {
+public interface CandidatoRepository extends PagingAndSortingRepository<Candidato, Integer> {
     
     Candidato findOneByIdCandidato(int idcandidato);
     Iterable<Candidato> findByStatus(String status);
